test(genetion): Moves generation tests in spoon/processing/CtGenertionTest
Closes #711
